+++ app/views/header.php
@@ -186,6 +186,9 @@ $navGroups = [
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-sRIl4kxILFvY47J16cr9ZwB07vP4J8+LH7qKQnuqkuIAvNWLzeN8tE5YBujZqJLB" crossorigin="anonymous">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css" rel="stylesheet">
     <link href="/assets/app.css" rel="stylesheet">
+    <?php if (function_exists('rich_text_editor_assets_requested') && rich_text_editor_assets_requested()): ?>
+        <link href='/assets/rich-text-editor.css' rel='stylesheet'>
+    <?php endif; ?>
 </head>
 
 <body>